amended Home interface names and moved to own file
