+++ src/pages/Home/Home.interfaces.ts
@@ -1,0 +1,13 @@
+export interface PartialApiData {
+  id: string;
+  image_url: string;
+  name: string;
+  description: string;
+}
+
+export interface DrinksListData {
+  id: string;
+  imageUrl: string;
+  name: string;
+  description: string;
+}
